+++ app/src/main/java/com/gianlu/pretendyourexyzzy/LoadingActivity.java
@@ -51,6 +51,7 @@ import com.google.firebase.auth.FirebaseUser;
 
 import org.jetbrains.annotations.NotNull;
 import org.json.JSONException;
+import org.json.JSONException;
 import org.json.JSONObject;
 
 import java.security.SecureRandom;