+++ app/src/main/java/com/gianlu/pretendyourexyzzy/api/urbandictionary/AutoCompleteResults.java
@@ -1,5 +1,7 @@
 package com.gianlu.pretendyourexyzzy.api.urbandictionary;
 
+import androidx.annotation.NonNull;
+
 import org.json.JSONArray;
 import org.json.JSONException;
 import org.json.JSONObject;
@@ -8,7 +10,7 @@ import java.util.ArrayList;
 
 public class AutoCompleteResults extends ArrayList<AutoCompleteResults.Item> {
 
-    AutoCompleteResults(JSONArray array) throws JSONException {
+    AutoCompleteResults(@NonNull JSONArray array) throws JSONException {
         super(array.length());
 
         for (int i = 0; i < array.length(); i++)
@@ -19,7 +21,7 @@ public class AutoCompleteResults extends ArrayList<AutoCompleteResults.Item> {
         public final String term;
         public final String preview;
 
-        Item(JSONObject obj) throws JSONException {
+        Item(@NonNull JSONObject obj) throws JSONException {
             term = obj.getString("term");
             preview = obj.getString("preview");
         }